(fix) separated routes from controllers on the backend
