+++ server/server.js
@@ -2,10 +2,10 @@ var express = require('express');
 var app = express();
 
 // middleware
-require('./config/middleware')(app, express);
+require('./config/middleware.js')(app, express);
 
 // routes
-require('./config/routes')(app);
+require('./config/routes.js')(app);
 
 var port = process.argv[2] || 8080;
 app.listen(port, function () {
